refactored and improved "save run".
- File menu entry "save run" now saves to the source file, while the new
entry "save run as" lets the user choose a file. If this file exists
already, the user has to choose whether to overwrite.
- initially the run is now saved to a temporary file in the same folder
as the output file (name scheme: ###<output filename>_<unique
numbers>.save_in_progress). This avoids corruption of the input file
when saving to the source file in case CLANS crashes or is force-closed
during the save operation. This is mainly important for large files or
when running CLANS remotely. Once all data has been written to the
temporary file, it is moved to its final destination. The drawback of
this approach is the temporarily double amount of required space during
the save operation.
- added hopefully helpful warning dialogs informing the user about
failed save operations
- the methods for saving runs in ClusterData now throw exceptions that
are handled by the calling code instead of trying to open windows
themselves