Prepare MinimalHsp so that it can be used as key in a HashMap
HashMap requires a key and a value, both must be Objects, that means both
are pointers in the HasMap in require 8 bytes each in 64 bit Java.

Additionally, comes the memory for the Objects, if we use the same object for
key and value, we can save that memory.

To achieve that we need be able to use MinimalHsp as key in a HashMap, since
we only want to use query and hit of MinimalHsp, the overriden methods
hashCode and equals should only depend on those.
And query and hit should be final so that they cannot be changed once 
MinimalHsp is in a HashMap, since this would screw up the HashMap.