Use MinimalAttractionValue in HashMaps as keys to itself instead of Strings
Strings need a lot of memory for representing two numbers seperated by an
underscore. However, the value for the key is already contained in the
MinimalAttractionValue object itself.

To use an MinimalAttractionValue as key to itsself, it quals and hashCode
function must depend on its values of query and hit. Since the att field is
supposed the value part in the HashTable, this is ignored by quals and hashCode.

This is a bit wired, but HashMap does not allow primitive types then a long
as key would be the choice or a pair of two ints, if Java would allow to pass
the object itself than a pointer onto it.