Replace HashMap by HashSet in ClusterDetectuin.java to reduce code complexity
This avoids adding dummy objects or fields. In principle, this could reduce memory
needs, however HashSet uses internally a HashMap and uses a static dummy Object for
filling the value part. That is a not very nice implementation.

However, this is now out of sight of the programmer so that other code issues get clearer.