Loop only through the attraction values that belong to the nodes of interests on ConvexClustering
In the worst case, each node has an attraction value to every other node. That are O(N²) attraction
avlues, if N is the number of nodes (aka sequences).  The old version looked for each node on all the
connections instead of just the connections of that particular node, which needed thus O(N²) loop
iterations. With the new implementation it just needs in the inner loop O(N) iteration. Which improves
the overall algorithm from O(N⁴) to O(N³).

This is a big improvement in speed. However, it cost O(N²) extra memory, which was however transiently
needed to load the data.