Fix averaging the attraction values in ClusterData.compute_attraction_values
The attraction values for the same edges seem to be supposed to be avaraged.
However, it was something else then avaraging.

If there was only an edge between node A and B but not between B and A, then 
the attraction value would be only half the size if it where.

In fact, it is questionable whether the attraction values should be treated differently
if they come from two different HSPs, then if they came from the same.